// src/interfaces/user_data/IUserProfileData.ts
export interface IUserProfileData {
  age: string;
  createdAt: string | null;
  gender: string | null;
  height: string | null;
  id: number | null;
  img: string | null;
  name: string | null;
  nickname: string | null;
  surname: string | null;
  updatedAt: string | null;
  userId: number | null;
  weight: string | null;
}